+++ finagle-exp/src/main/scala/com/twitter/finagle/exp/DarkTrafficFilter.scala
@@ -1,7 +1,10 @@
 package com.twitter.finagle.exp
 
+import com.twitter.finagle._
 import com.twitter.finagle.stats.StatsReceiver
-import com.twitter.finagle.{Service, SimpleFilter}
+import com.twitter.finagle.tracing.Annotation.BinaryAnnotation
+import com.twitter.finagle.tracing.ForwardAnnotation
+import com.twitter.finagle.util.Rng
 import com.twitter.logging.{HasLogLevel, Level, Logger}
 import com.twitter.util.Future
 
@@ -22,7 +25,7 @@ class DarkTrafficFilter[Req, Rep](
     extends SimpleFilter[Req, Rep]
     with AbstractDarkTrafficFilter {
 
-  import DarkTrafficFilter.log
+  import DarkTrafficFilter._
 
   def this(
     darkService: Service[Req, Rep],
@@ -31,12 +34,16 @@ class DarkTrafficFilter[Req, Rep](
   ) = this(darkService, enableSampling, statsReceiver, false)
 
   def apply(request: Req, service: Service[Req, Rep]): Future[Rep] = {
-    if (forwardAfterService) {
-      service(request).ensure {
-        sendDarkRequest(request)(enableSampling, darkService)
+    // Set an identifier so we can later determine which two requests
+    // match, as well as which request is the dark one.
+    ForwardAnnotation.let(newKeyAnnotation()) {
+      if (forwardAfterService) {
+        service(request).ensure {
+          sendDarkRequest(request)(enableSampling, darkService)
+        }
+      } else {
+        serviceConcurrently(service, request)(enableSampling, darkService)
       }
-    } else {
-      serviceConcurrently(service, request)(enableSampling, darkService)
     }
   }
 
@@ -49,6 +56,10 @@ class DarkTrafficFilter[Req, Rep](
   }
 }
 
-private object DarkTrafficFilter {
+object DarkTrafficFilter {
   val log: Logger = Logger.get("DarkTrafficFilter")
+
+  val DarkRequestAnnotation = BinaryAnnotation("clnt/dark_request", true)
+  def newKeyAnnotation() =
+    BinaryAnnotation("clnt/dark_request_key", Rng.threadLocal.nextLong(Long.MaxValue))
 }